services/api-gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "1.0.0"
description = "API Gateway — unified entry point for LLM Management Platform"
authors = ["Seungha Lee <lesh915@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.32"}
httpx = "^0.27"
pydantic = "^2.0"
python-jose = {extras = ["cryptography"], version = "^3.3"}
python-multipart = "^0.0.12"

[tool.poetry.dev-dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
pytest-mock = "^3.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
